requirements.txt
aiohttp>=3.9.0
eth-account>=0.11.0
numpy>=1.26.0
scipy>=1.12.0
web3>=6.15.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aioresponses>=0.7.6
